requirements.txt
--no-binary=opencc
numpy<2.0
scipy
librosa==0.10.2
numba
gradio<5
ffmpeg-python
onnxruntime
tqdm
cn2an
pypinyin
pyopenjtalk>=0.4.1
g2p_en
torchaudio
transformers>=4.43,<=4.50
peft<0.18.0
chardet
PyYAML
psutil
jieba_fast
jieba
split-lang
fast_langdetect>=0.3.1
wordsegment
ToJyutping
g2pk2
ko_pron
opencc
python_mecab_ko; sys_platform != 'win32'
fastapi[standard]>=0.115.2
starlette<1.0
x_transformers
torchmetrics<=1.5
pydantic<=2.10.6
soundfile
